refactor(ui): move DiagnosticSummaryView and ErrorSection to UI/Components\n\n- No logic changes; preserve accessibility and styling\n- Core tests green via ./run-core-tests.sh
