+++ Sources/KeyPath/UI/ContentView.swift
@@ -671,72 +671,7 @@ struct ContentView: View {
 
 // RecordingSection moved to UI/Components/RecordingSection.swift
 
-struct ErrorSection: View {
-    @ObservedObject var kanataManager: KanataViewModel // Phase 4: MVVM
-    @Binding var showingInstallationWizard: Bool
-    let error: String
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack {
-                Image(systemName: "exclamationmark.triangle.fill")
-                    .foregroundColor(.orange)
-                    .font(.headline)
-
-                Text("KeyPath Error")
-                    .font(.headline)
-                    .foregroundColor(.primary)
-
-                Spacer()
-
-                Button("Fix Issues") {
-                    Task {
-                        AppLogger.shared.log(
-                            "🔄 [UI] Fix Issues button clicked - attempting to fix configuration and restart")
-
-                        // Create a default user config if missing
-                        let created = await kanataManager.createDefaultUserConfigIfMissing()
-
-                        if created {
-                            await MainActor.run {
-                                kanataManager.lastError = nil
-                            }
-                            AppLogger.shared.log(
-                                "✅ [UI] Created default config at ~/Library/Application Support/KeyPath/keypath.kbd"
-                            )
-                        } else {
-                            // Still not fixed – open wizard to guide the user
-                            showingInstallationWizard = true
-                        }
-
-                        // Try starting after config creation
-                        await kanataManager.startKanata()
-                        await kanataManager.updateStatus()
-                        AppLogger.shared.log(
-                            "🔄 [UI] Fix Issues completed - service status: \(kanataManager.isRunning)")
-                    }
-                    // Only open wizard above if needed
-                }
-                .buttonStyle(.borderedProminent)
-                .controlSize(.small)
-            }
-
-            Text(error)
-                .font(.body)
-                .foregroundColor(.secondary)
-                .fixedSize(horizontal: false, vertical: true)
-        }
-        .padding()
-        .background(
-            RoundedRectangle(cornerRadius: 12)
-                .fill(Color(NSColor.textBackgroundColor).opacity(0.95))
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 12)
-                .stroke(Color.orange.opacity(0.35), lineWidth: 1)
-        )
-    }
-}
+// ErrorSection moved to UI/Components/ErrorSection.swift
 
 struct DiagnosticSummarySection: View {
     let criticalIssues: [KanataDiagnostic]
@@ -806,79 +741,7 @@ struct DiagnosticSummarySection: View {
 
 // StatusMessageView moved to UI/Components/StatusMessageView.swift
 
-struct DiagnosticSummaryView: View {
-    let criticalIssues: [KanataDiagnostic]
-    let onViewDiagnostics: () -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            HStack {
-                Image(systemName: "exclamationmark.triangle.fill")
-                    .foregroundColor(.orange)
-                    .font(.title2)
-
-                VStack(alignment: .leading, spacing: 4) {
-                    Text("System Issues Detected")
-                        .font(.headline)
-                        .foregroundColor(.primary)
-
-                    Text("\(criticalIssues.count) critical issue\(criticalIssues.count == 1 ? "" : "s") need attention")
-                        .font(.subheadline)
-                        .foregroundColor(.secondary)
-                }
-
-                Spacer()
-
-                Button(action: onViewDiagnostics, label: {
-                    Text("View Details")
-                        .font(.subheadline)
-                        .foregroundColor(.blue)
-                })
-                .buttonStyle(.plain)
-            }
-
-            // Show first 2 critical issues as preview
-            ForEach(Array(criticalIssues.prefix(2).enumerated()), id: \.offset) { _, issue in
-                HStack(alignment: .top, spacing: 8) {
-                    Image(systemName: issue.severity == .critical ? "exclamationmark.circle.fill" : "xmark.circle.fill")
-                        .foregroundColor(issue.severity == .critical ? .red : .orange)
-                        .font(.caption)
-                        .padding(.top, 2)
-
-                    VStack(alignment: .leading, spacing: 2) {
-                        Text(issue.title)
-                            .font(.caption)
-                            .fontWeight(.medium)
-                            .foregroundColor(.primary)
-
-                        if !issue.description.isEmpty {
-                            Text(issue.description)
-                                .font(.caption2)
-                                .foregroundColor(.secondary)
-                                .fixedSize(horizontal: false, vertical: true)
-                        }
-                    }
-                }
-            }
-
-            if criticalIssues.count > 2 {
-                Text("... and \(criticalIssues.count - 2) more issue\(criticalIssues.count - 2 == 1 ? "" : "s")")
-                    .font(.caption)
-                    .foregroundColor(.secondary)
-                    .padding(.leading, 16)
-            }
-        }
-        .padding(12)
-        .background(
-            RoundedRectangle(cornerRadius: 8)
-                .fill(Color.orange.opacity(0.1))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 8)
-                        .stroke(Color.orange.opacity(0.3), lineWidth: 1)
-                )
-        )
-    }
-}
+// DiagnosticSummaryView moved to UI/Components/DiagnosticSummaryView.swift
 
 #Preview {
     let manager = KanataManager()
